playground/web-api/requirements.txt
django
djangorestframework
drf-spectacular
pytest